# Conflicts:
#	database_setup.pyc

+++ catalog_app.py
@@ -1,5 +1,5 @@
 import os
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for, flash, jsonify
 
 appCatalog = Flask(__name__)
 import cgi
@@ -7,6 +7,7 @@ from sqlalchemy import create_engine, func
 from sqlalchemy.orm import sessionmaker, aliased
 from database_setup import Base, CategoryTable, ItemTable
 from werkzeug import secure_filename
+#Import all the necesary files and clases for authentication
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
 import httplib2
@@ -14,10 +15,10 @@ import json
 from flask import make_response
 import requests
 
-appCatalog.config['SESSION_TYPE'] = 'filesystem'
 from flask import session as login_session
 import random, string
 
+appCatalog.config['SESSION_TYPE'] = 'filesystem'
 CLIENT_ID = json.loads(open('client_secret.json','r').read())['web']['client_id']
 engine = create_engine ('sqlite:///ItemCatalog.db')
 Base.metadata.bind = engine
@@ -26,8 +27,10 @@ appCatalog.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+#Methods and views related to the user authentication.
 @appCatalog.route('/gconnect',methods=['POST'])
 def gconnect():
+	#Gconnect function for linking the gmail account with the user session
 	if request.args.get('state') != login_session['state']:
 		response = make_response(json.dumps('Invalid state parameter.'),401)
 		response.headers['Content-Type'] = 'application/json'
@@ -41,11 +44,11 @@ def gconnect():
 		response = make_response(json.dumps('Failed to upgrade the authorization code.'),401)
 		response.headers['Content-Type'] = 'application/json'
 		return response
+	#Get the tokens for the current session
 	token = credentials.access_token
-
 	url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'%token)
 	h = httplib2.Http()
-	
+	#Get the result of the google api
 	result = json.loads(h.request(url, 'GET')[1])
 
 	#If there was an error in the access token info, kill it.
@@ -67,61 +70,117 @@ def gconnect():
 	params = {'access_token':credentials.access_token, 'alt':'json'}
 	answer = requests.get(userinfo_url, params=params)
 	data = json.loads(answer.text)
-	#COLOCAR CODIGO PARA RETORNAR UNA VEZ SE REALIZE LA VALIDACION DE TODO.
+	login_session['name'] = data['name']
+	login_session['picture'] = data['picture']
+	response = ''
+	response += 'Connecting....'
+	#flash('Welcome %s'%login_session['name'])
+	return response
+
+@appCatalog.route('/gdisconnect')
+def gdisconnect():
+	#Only disconnect a connected user
+	credentials = login_session.get('credentials')
+	if credentials is None:
+		response = make_response(json.dumps('Current user is not connected.'),401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+	#Execute Get HTTP request to revoke current token
+	access_token = credentials.access_token
+	url = 'https://accounts.google.com/o/oauth2/revoke?token=%s'%access_token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[0]
+	#If succes delete al login session variables
+	print result
+	if result['status'] == '200':
+		del login_session['credentials']
+		del login_session['gplus_id']
+		del login_session['name']
+		del login_session['picture']
+
+		response = make_response(json.dumps('Succesfully disconnected!.'))
+		response.headers['Content-Type'] = 'application/json'
+		return redirect('/')
+
+	return 'Response status: '+result['status']
 
 @appCatalog.route('/')
 @appCatalog.route('/login')
 def showLogin():
+	"""Login view"""
+	#First create the corresponding state for the login session.
 	state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in range(32))
 	login_session['state'] = state
-
 	sports_items = session.query(CategoryTable.name,CategoryTable.id,ItemTable.name).join(ItemTable.category).group_by(CategoryTable.id).order_by(CategoryTable.name).all()
 
 	sport_list = []
 	for sport in sports_items:
 		sport_list.append({'name':str(sport[0]),'idcategory':sport[1],'item':sport[2]})
-		
-	return render_template('home.html',sport_list=sport_list,item_title='Latest Items',buttons_rows='',STATE=state)
+	#Get all categories and get the latest item.
+	return render_template('home.html',sport_list=sport_list,item_title='Latest Items',buttons_rows=False,STATE=state)
+
 
 @appCatalog.route('/home')
 def home():
 	sports_items = session.query(CategoryTable.name,CategoryTable.id,ItemTable.name).join(ItemTable.category).group_by(CategoryTable.id).order_by(CategoryTable.name).all()
+	if 'name' not in login_session:
+		return redirect('/')
 
+	name = login_session['name']
+	image = login_session['picture']
 	sport_list = []
 	for sport in sports_items:
 		sport_list.append({'name':str(sport[0]),'idcategory':sport[1],'item':sport[2]})
-	buttons_rows = '<a class="btn btn-primary" href="/add-item" role="button">Create Item</a>'
-	buttons_rows +='<a class="btn btn-primary" href="/add-category" role="button">Create Category</a>'
-	return render_template('home.html',sport_list=sport_list,item_title='Latest Items',buttons_rows=buttons_rows)
+	print login_session
+	
+	return render_template('home.html',sport_list=sport_list,item_title='Latest Items',buttons_rows=True,username=name,image=image)
 
 @appCatalog.route('/filter-items/<int:idcategory>')
 def filterItem(idcategory):
 	sport_items = session.query(CategoryTable.name,ItemTable.id,ItemTable.name).join(ItemTable.category).filter(CategoryTable.id == idcategory).all()
 	test = session.query(CategoryTable).all()
 	items_list = []
+	name = ''
+	image = ''
+	show_rows = True
+	if 'name' not in login_session:
+		show_rows = False
+	else:
+		name = login_session['name']
+		image = login_session['picture']
 
 	for item in sport_items:
 		category_name = item[0]
 		items_list.append({'category_name':str(item[0]),'id_item':item[1],'item_name':item[2]})
 	info = "In this section you can view all items related to a specific category. Also you can "
 	info += "filter results in the text field by typing the item's name."
-	return render_template('items-by-category.html',items_list=items_list,category_name=category_name,info=info)
+
+	return render_template('items-by-category.html',items_list=items_list,category_name=category_name,info=info, show_rows=show_rows, username=name, image=image)
 
 @appCatalog.route('/add-item')
 def addItem():
 	sports = session.query(CategoryTable).all()
 	sport_list = []
+	if 'name' not in login_session:
+		return redirect('/')
+
+	name = login_session['name']
+	image = login_session['picture']
 	info = "Here you can create new items, with a description and asociate them to a specific category."
 	for sport in sports:
 		sport_list.append({'name':str(sport.name),'id':sport.id})
-	return render_template('create-item.html',sport_list=sport_list, info=info)
+	return render_template('create-item.html',sport_list=sport_list, info=info, username=name, image=image)
 
 @appCatalog.route('/edit-item/<int:iditem>')
 def editItem(iditem):
 	model = session.query(ItemTable).filter(ItemTable.id == iditem).one()
 	sports = session.query(CategoryTable).all()
 	info = "Here you can edit the selected item, also the description and the category asociated with it."
+	if 'name' not in login_session:
+		return redirect('/')
 
+	name = login_session['name']
+	image = login_session['picture']
 	sport_list = []
 	item_model = []
 	for sport in sports:
@@ -129,7 +188,7 @@ def editItem(iditem):
 
 	item_model.append({'name':model.name, 'id':iditem,'category':model.category_id,'image':model.image,'description':model.description,'sport_list':sport_list})
 
-	return render_template('update-item.html',item_model=item_model, info=info)
+	return render_template('update-item.html',item_model=item_model, info=info, username=name, image=image)
 
 @appCatalog.route('/update-item/<int:iditem>',methods=['GET','POST'])
 def updateItem(iditem):
@@ -153,9 +212,13 @@ def updateItem(iditem):
 
 @appCatalog.route('/add-category')
 def addCategory():
+	if 'name' not in login_session:
+		return redirect('/')
 
+	name = login_session['name']
+	image = login_session['picture']
 	info = "In this section you'll be able to create a new category."
-	return render_template('create-category.html',info=info)
+	return render_template('create-category.html',info=info, username=name, image=image)
 
 @appCatalog.route('/insert-item',methods=['GET','POST'])
 def createItem():
@@ -188,6 +251,8 @@ def createCategory():
 
 @appCatalog.route('/remove-item/<int:iditem>',methods=['GET','POST'])
 def removeItem(iditem):
+	if 'name' not in login_session:
+		return redirect('/')
 
 	item = session.query(ItemTable).filter_by( id = iditem ).one()
 	session.delete(item)
@@ -199,6 +264,16 @@ def viewItem(iditem):
 	info = "Here you can view the information about the item."
 	model = session.query(ItemTable).filter(ItemTable.id == iditem).one()
 	sports = session.query(CategoryTable).all()
+	show_user = True
+	name = ''
+	image = ''
+
+	if 'name' not in login_session:
+		show_user = False
+	else:
+		name = login_session['name']
+		image = login_session['picture']
+
 	category = ''
 	item_model = []
 	for sport in sports:
@@ -206,8 +281,8 @@ def viewItem(iditem):
 			category = str(sport.name)
 
 	item_model.append({'name':model.name, 'id':iditem,'category':category, 'description':model.description, 'image':model.image})
+	return render_template('view-item.html',item_model=item_model, info = info, show_user=show_user, username=name, image=image)
 
-	return render_template('view-item.html',item_model=item_model, info = info)
 if __name__=='__main__':
 	appCatalog.debug = True;
 	appCatalog.secret_key = 'JsBwnLvJQPsopQ0UN6Zkf9Av'
